fix: keep `lsp_changes` for each server
This is needed as the notifications could be sent at different times for each server.

For example we have:
```
lsp_changes = { change_1, change2, change3 }
```
And the change notification is queued for server A and server B.

Server A is free, so the notification is sent quickly, and `lsp_changes` is reset.
Server B still has the notification in the queue.

In the meantime, a new change is added:
```
lsp_changes = { change_4 }
```
And the notification is queued for A and B.
But, because we send the notification with `overwrite` set, B overwrites the "old" change notification with this new one, thus losing the old changes.